app/pyproject.toml
[project]
name = "podcast-automator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<4.0"
readme = "README.md"
authors = [
  { name = "sunabalog" },
]
dependencies = [
    "boto3>=1.42.17",
    "feedgen>=1.0.0",
    "feedparser>=6.0.12",
    "functions-framework==3.*",
    "google-cloud-aiplatform>=1.132.0",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-storage>=3.7.0",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
  "pytest-cov>=4.1.0",
  "pytest>=7.4.3",
  "ruff>=0.1.6",
]

[tool.pytest.ini_options]
addopts = "-vvv --durations=3"
testpaths = ["tests"]
pythonpath = [".", "src"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
pydocstyle.convention = "google"
select = [
  "ALL",
]
ignore = [
  "ISC001",  # single-line-implicit-string-concatenation produces conflicts
  "ANN003",  # kwargs without type check
  "COM812",  # conflicts when used with the formatter
  "TRY300",  # do not enforce try-else
  "PERF203",  # allow try except in loop
  "TRY301",  # allow raise in try
  "TRY003",  # allow long exception messages
  "EM101",  # allow string literals in exceptions
  "ARG001",  # allow unused arguments (Lambda context, fixtures)
  "ANN401",  # allow Any type for Lambda context
]
unfixable = [
  "S",  # do not auto fix security related findings
]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = [
  "S",  # ignore security related warnings in tests
  "C",  # ignore various C categories in tests
  "D1",  # ignore undocumented code in tests
  "ANN",  # ignore some type annotation checks in tests
  "PLR2004",  # ignore magic number compare
  "SLF001",  # allow private member access for cache reset
]

[tool.ruff.lint.mccabe]
max-complexity = 20

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv.workspace]
members = ["."]